Create SIP_TEMPLATE
Starting a NEW SIP Template for easy of SIP Author to start somewhere which will output consistent formatting. 

Support needed to refine instruction:
* Wording the License field part better + the License and Copyright Section. 

Open questions:
1. In BIP process, they recently changed "created" field in Preamble to "Assigned" https://github.com/bitcoin/bips/pull/1970 
we could use this opportunity to update to align. I asked the BIP Editor Jon, his reasoning was "I suppose it's simply to provide a date for people to see when a BIP was launched. A number assignment seems to be the most interesting date for people, more than when the draft PR was opened or merged."